requirements.txt
Django==2.2.10
pytz==2019.2
sqlparse==0.3.0
psycopg2-binary
django-crispy-forms==1.7.2
django-debug-toolbar==2.0
celery==4.3.0
flower==0.9.3
setuptools==41.4.0
django-extended-choices==1.3.3
django-range-slider==0.2.4
django-environ
boto3
django-aws-xray
requests
validators
watchtower
